// enums/signup.enum.ts
export const enum ESignupLocator {
    EmailInput = 'EMAIL_INPUT',
    FirstNameInput = 'FIRSTNAME_INPUT',
    LastNameInput = 'LASTNAME_INPUT',
    PasswordInput = 'PASSWORD_INPUT',
    PolicyCheckbox = 'POLICY_CHECKBOX',
    SignUpButton = 'SIGNUP_BUTTON',
    ErrorModal = "ERROR_MODAL",
    OtpInput = "OTP_INPUT",
    OtpErrorMessage = "OTP_ERROR_MESSAGE",
    SuccessModal= "SUCCESS_MODAL",
    ResendButton = "RESEND_BUTTON",
    BackToSignInButton = "BACK_TO_SIGNIN_BUTTON"
}